backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.3.4
python-multipart==0.0.9
redis==5.0.7
celery==5.4.0
httpx==0.27.0
orjson==3.10.6
python-dateutil==2.9.0.post0
requests==2.32.3
openai==1.35.10
google-generativeai==0.7.2
playwright==1.46.0
beautifulsoup4==4.12.3
lxml==5.2.2
